pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "echovault"
version = "0.3.0"
description = "Local-first memory system for coding agents with Obsidian-native markdown storage"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "sqlite-vec>=0.1.6",
    "pysqlite3>=0.6.0",
    "httpx>=0.27.0",
    "mcp>=1.0",
    "textual>=0.52.1",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
memory = "memory.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
